☀️ 53% – Adopt translation units with holes as the standard shape
Pret's discipline at function granularity: main-image work happens
inside declared translation units — address-ordered include composites
holding every owner in a range, with exact owners as production
sources, in-progress owners as candidate drafts, and everything else as
retained-assembly holes listed in the manifest. Holes are whole
functions, never asm() inside a C body, and DONE never counts a hole.

tools/unit-scaffold generates the composite and manifest entry for a
range; the recovering agent resolves declaration collisions and scores
with candidate-show --unit until every previously exact owner is exact
again. CONTRIBUTING documents the method and folds it into the recovery
flow.

The pilot proves the mechanism and its value in one stroke. The
battle-summon unit (0x080b7424..0x080b770c) exposed that two exact
owners called Func_08077008 through different prototypes; unifying on
the s32 argument kept both byte-exact — the u16 guess forces a caller-
side mask and regresses — so the unit recovered the true shared
declaration while candidates 080b7548 and 080b75dc now score inside
their authentic context.

+++ games/gs1/src/battle/summon/layout_positions.c
@@ -5,7 +5,7 @@ struct BattleActorDefinition {
     u8 class_id;
 };
 
-struct BattleActorDefinition *Func_08077008(u16 actor_id);
+struct BattleActorDefinition *Func_08077008(s32 actor_id);
 s32 Func_080c23c0(u8 class_id);
 
 #define LayoutSummonPositions Func_080b7424
